util: extract test that should only be ran with -race enabled and unskip it
